fill in docstrings, close #3
The biggest part of the documentation text is included
in the docstrings in discid.py
Only a few connectors and automatic placeholders are in the rst files.

Python objects (discid and python stdlib) are linked in the
docstrings.
Especially for "str" it is important to link to the correct python
version. (Python 3 in this case)

If _get_default_device is mocked (so building the doc doesn't need
libdiscid), then it is explicitely set to None
so the docstring documentation works.

Pylint doesn't expect docstrings for module constants and attributes,
so it shows these as warnings.
Other pylint warnings should be gone.