add (major) version argument to _find_library
Not really sure if libdiscid will ever reach 1.0.0 stable,
but now we don't have "magic" numbers in _find_library and it is
completely generic and could possibly be used by other tools (in code).

It is still a private function because exporting a nice "find_library"
is not in the scope of this module and messes the API up.