don't request features we don't know of
I removed the possibility to request all features (UINT_MAX) again.
It would lead to weird behavior as phw stated.

Features that wouldn't need additional changes in this module,
other than adding the mapping should be rare and python-discid,
would get added to this module fast anyways.
It is more likely that libdiscid is older than python-discid since it is
in stable repositories for Linux and needs new builds on other
platforms.