move constant "docstrings" to __init__.py
This way sphinx can access them.
Either way sphinx thinks they are defined in the discid package,
and tries to extract the "docstring" from there.
Actually the "docstring" is just sphinx reading the source in that case.

Another option would be to access them as disid.libdiscid.* in sphinx,
but that wouldn't document them as being in the main package.